+++ lib/ui/groups/GroupDetailPanel.dart
@@ -347,9 +347,18 @@ class _GroupDetailPanelState extends State<GroupDetailPanel> implements Notifica
         // Clear _postId in order not to redirect on the next group load.
         _postId = null;
         if (post != null) {
-          Navigator.push(context, CupertinoPageRoute(builder: (context) => GroupPostDetailPanel(group: _group, post: post)));
+          String? redirectId = post.topParentId ?? post.id; //Get mainPostId. if topParentId == null then this is the main Post
+          if(StringUtils.isNotEmpty(redirectId)){
+            Groups().loadGroupPost(groupId: _group!.id, postId: redirectId).then((mainPost) {
+              _decreaseProgress();
+              Navigator.push(context, CupertinoPageRoute(builder: (context) => GroupPostDetailPanel(group: _group, post: mainPost)));
+            });
+          } else {
+            _decreaseProgress();
+          }
+        } else {
+          _decreaseProgress();
         }
-        _decreaseProgress();
       });
     }
   }